refactor: split badge generation into standalone script
Extract badge generation from regen_requirements.py into a new
scripts/regen_badges.py so the dependabot-regen workflow can call
it independently (without invoking uv lock). Add pre-commit hook
for badge regeneration. Update the dependabot workflow to regenerate
badges alongside requirements on Dependabot PRs.
