+++ .github/workflows/dependabot-regen.yml
@@ -2,7 +2,8 @@
 #
 # Dependabot (uv ecosystem) updates pyproject.toml + uv.lock but cannot
 # run post-update commands.  This workflow regenerates the derived
-# requirements*.txt files and commits them back to the PR branch.
+# requirements*.txt files and badge JSON files, then commits them back
+# to the PR branch.
 #
 # Uses GITHUB_TOKEN to push.  If the pushed commit does not retrigger CI
 # (a known GitHub limitation with GITHUB_TOKEN), switch to a GitHub App
@@ -25,7 +26,7 @@ concurrency:
 
 jobs:
   regen:
-    name: Regenerate requirements
+    name: Regenerate requirements and badges
     if: github.actor == 'dependabot[bot]'
     runs-on: ubuntu-latest
     timeout-minutes: 10
@@ -57,14 +58,17 @@ jobs:
           uv export --format requirements-txt --no-emit-project \
             --output-file requirements-dev.txt
 
+      - name: Regenerate badge JSON files
+        run: uv run python scripts/regen_badges.py
+
       - name: Commit and push if changed
         run: |
-          if git diff --quiet requirements.txt requirements-dev.txt; then
-            echo "Requirements files are up to date — nothing to commit."
+          if git diff --quiet requirements.txt requirements-dev.txt badges/; then
+            echo "Derived files are up to date — nothing to commit."
             exit 0
           fi
           git config user.name "github-actions[bot]"
           git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git add requirements.txt requirements-dev.txt
-          git commit -m "chore(deps): regenerate requirements from uv.lock"
+          git add requirements.txt requirements-dev.txt badges/
+          git commit -m "chore(deps): regenerate requirements and badges from uv.lock"
           git push
