+++ badges/ruff.json
@@ -1,6 +1,6 @@
 {
   "schemaVersion": 1,
   "label": "ruff",
-  "message": "v0.15.9",
+  "message": "v0.15.10",
   "color": "D7FF64"
 }
