+++ badges/ty.json
@@ -1,6 +1,6 @@
 {
   "schemaVersion": 1,
   "label": "ty",
-  "message": "v0.0.29",
+  "message": "v0.0.31",
   "color": "D7FF64"
 }
